requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/dev.in
#
aiohttp==3.8.5
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
asttokens==2.4.0
    # via stack-data
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
backcall==0.2.0
    # via ipython
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   -c requirements/prod.txt
    #   kfp
comm==0.1.4
    # via ipykernel
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via
    #   gcsfs
    #   ipython
docstring-parser==0.15
    # via kfp
executing==2.0.0
    # via stack-data
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.9.2
    # via gcsfs
gcsfs==2023.9.2
    # via -r requirements/dev.in
google-api-core[grpc]==2.10.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-pipeline-components
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   kfp
google-auth==2.23.2
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
    #   kubernetes
google-auth-oauthlib==1.1.0
    # via gcsfs
google-cloud-aiplatform==1.33.1
    # via
    #   -r requirements/dev.in
    #   google-cloud-pipeline-components
google-cloud-bigquery==3.11.4
    # via
    #   -r requirements/dev.in
    #   google-cloud-aiplatform
google-cloud-core==2.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-pipeline-components==2.4.0
    # via -r requirements/dev.in
google-cloud-resource-manager==1.6.3
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.12.6
    # via -r requirements/dev.in
google-cloud-storage==2.11.0
    # via
    #   -r requirements/dev.in
    #   gcsfs
    #   google-cloud-aiplatform
    #   kfp
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.6.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.60.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
grpcio==1.59.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.48.2
    # via google-api-core
idna==3.4
    # via
    #   -c requirements/prod.txt
    #   requests
    #   yarl
ipykernel==6.25.2
    # via -r requirements/dev.in
ipython==8.16.0
    # via ipykernel
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via
    #   -c requirements/prod.txt
    #   google-cloud-pipeline-components
joblib==1.3.2
    # via scikit-learn
jupyter-client==8.3.1
    # via ipykernel
jupyter-core==5.3.2
    # via
    #   ipykernel
    #   jupyter-client
kfp==2.3.0
    # via
    #   -r requirements/dev.in
    #   google-cloud-pipeline-components
kfp-pipeline-spec==0.2.2
    # via kfp
kfp-server-api==2.0.1
    # via kfp
kubernetes==26.1.0
    # via kfp
markupsafe==2.1.3
    # via
    #   -c requirements/prod.txt
    #   jinja2
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.5.8
    # via ipykernel
numpy==1.26.0
    # via
    #   pandas
    #   scikit-learn
    #   scipy
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   ipykernel
pandas==2.1.1
    # via -r requirements/dev.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==3.10.0
    # via jupyter-core
prompt-toolkit==3.0.39
    # via ipython
proto-plus==1.22.3
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
protobuf==3.20.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   kfp
    #   kfp-pipeline-spec
    #   proto-plus
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pygments==2.16.1
    # via ipython
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   jupyter-client
    #   kfp-server-api
    #   kubernetes
    #   pandas
python-dotenv==1.0.0
    # via -r requirements/dev.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   kfp
    #   kubernetes
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.31.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   kubernetes
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   kubernetes
requests-toolbelt==0.10.1
    # via kfp
rsa==4.9
    # via google-auth
scikit-learn==1.3.1
    # via -r requirements/dev.in
scipy==1.11.3
    # via scikit-learn
shapely==1.8.5.post1
    # via google-cloud-aiplatform
six==1.16.0
    # via
    #   asttokens
    #   kfp-server-api
    #   kubernetes
    #   python-dateutil
stack-data==0.6.3
    # via ipython
tabulate==0.9.0
    # via kfp
threadpoolctl==3.2.0
    # via scikit-learn
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.10.1
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   kfp
    #   kfp-server-api
    #   kubernetes
    #   requests
wcwidth==0.2.8
    # via prompt-toolkit
websocket-client==1.6.3
    # via kubernetes
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
